Feat: Display game type based on queue id
